pyproject.toml
[project]
name = "rces"
version = "0.1.0"
description = "Rust Command Execution Server"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0"
]
